move to external module dashboard
