--- src/main/java/uk/gov/hmcts/reform/civil/repositories/citizenui/Example.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.civil.repositories.citizenui;
-
-public class Example {
-}
